Update bean.js
You could ping everyone and that's bad

+++ src/commands/Fun/bean.js
@@ -25,8 +25,10 @@ module.exports = {
             return message.channel.send(`<@!${args[0]}> was beaned!`);
 
         } else {
+            let input = args.join(' ');
+            if ([/@everyone/, /@here/, /<@&\d{18}>/].some(e => e.test(args[0]))) return message.channel.send('i\'m not going to ping everyone and instead make you look like an idiot lmao!!! point and laugh at this dude');
             if (message.deletable) message.delete();
-            return message.channel.send(`${args.join(' ')} was beaned!`);
+            return message.channel.send(`${input} was beaned!`);
         }
     }
 };
